LP-834: Add menu fragment
