+++ VeryfiLensAndroidNetDemo/VeryfiLensAndroidNetDemo/MainActivity.cs
@@ -5,6 +5,7 @@ using AndroidX.ConstraintLayout.Widget;
 using Com.Veryfi.Lens;
 using Com.Veryfi.Lens.Helpers;
 using Org.Json;
+using VeryfiLensAndroidNetDemo.fragments;
 
 namespace VeryfiLensAndroidNetDemo;
 
@@ -25,77 +26,12 @@ public class MainActivity : AppCompatActivity
         SetUpVeryfiLens();
         SetUpVeryfiLensDelegate();
 
-        ListenerEvents();
+        var transaction = SupportFragmentManager.BeginTransaction();
+        transaction.Replace(Resource.Id.fragment_container, new MenuFragment());
+        transaction.Commit();
 
-    }
-
-    private void ListenerEvents()
-    {
-        var btnReceipts = FindViewById<ConstraintLayout>(Resource.Id.btn_receipts);
-        if (btnReceipts != null)
-        {
-            btnReceipts.Click += FabOnClick;
-        }
-
-        var btnLongReceipts = FindViewById<ConstraintLayout>(Resource.Id.btn_long_receipts);
-        if (btnLongReceipts != null)
-        {
-            btnLongReceipts.Click += FabOnClick;
-        }
-
-        var btnChecks = FindViewById<ConstraintLayout>(Resource.Id.btn_checks);
-        if (btnChecks != null)
-        {
-            btnChecks.Click += FabOnClick;
-        }
-
-        var btnCreditCards = FindViewById<ConstraintLayout>(Resource.Id.btn_credit_cards);
-        if (btnCreditCards != null)
-        {
-            btnCreditCards.Click += FabOnClick;
-        }
-
-        var btnBusinessCards = FindViewById<ConstraintLayout>(Resource.Id.btn_business_cards);
-        if (btnBusinessCards != null)
-        {
-            btnBusinessCards.Click += FabOnClick;
-        }
-
-        var btnOcr = FindViewById<ConstraintLayout>(Resource.Id.btn_ocr);
-        if (btnOcr != null)
-        {
-            btnOcr.Click += FabOnClick;
-        }
-
-        var btnW2 = FindViewById<ConstraintLayout>(Resource.Id.btn_w2);
-        if (btnW2 != null)
-        {
-            btnW2.Click += FabOnClick;
-        }
-
-        var btnW9 = FindViewById<ConstraintLayout>(Resource.Id.btn_w9);
-        if (btnW9 != null)
-        {
-            btnW9.Click += FabOnClick;
-        }
-
-        var btnBankStatements = FindViewById<ConstraintLayout>(Resource.Id.btn_bank_statements);
-        if (btnBankStatements != null)
-        {
-            btnBankStatements.Click += FabOnClick;
-        }
-
-        var btnBarcode = FindViewById<ConstraintLayout>(Resource.Id.btn_barcode);
-        if (btnBarcode != null)
-        {
-            btnBarcode.Click += FabOnClick;
-        }
     }
     
-    private void FabOnClick(object sender, EventArgs eventArgs)
-    {
-        VeryfiLens.ShowCamera();
-    }
 
     private void SetUpVeryfiLensDelegate()
     {
